updated search with contexts

+++ app/src/main/java/com/me/restaurantsmartsearch/fragment/ListFragment.java
@@ -178,7 +178,7 @@ public class ListFragment extends BaseFragment {
     public void searchOnline() {
         isSearchOnline = true;
         hideSoftKeyboard();
-        long lon = 0,lat = 0;
+        double lon = 0,lat = 0;
         String fields[] = {"name"};
         SearchAsyncTask searchAsyncTask = new SearchAsyncTask(AccentRemover.removeAccent(edSearch.getText().toString()),lon,lat,fields, new SearchAsyncTask.OnSearchComplete() {
             @Override
